Fix bug in handling programs that use '\r'
If the spawned program shows progress using a lines that only end
in \r instead of \n, the buffer would fill up and the spawned
program would die because stdout/stderr would fill up. This fixes
that by checking for both \r and \n.

There was also a typo in that the wrong length was used when
creating the buffer of data to copy from stdout/stderr